use sqlite instead of csv, full migrate later

+++ src/index.ts
@@ -1,7 +1,10 @@
 import fs from 'fs';
 import discord from 'discord.js'
+import { downloadAllAttachments, downloadAllStickers, downloadAllUsers } from './full-archive';
+import sqlite3 from 'sqlite3';
+import { open } from 'sqlite';
 
-const { token, guildid } = JSON.parse(fs.readFileSync('../.creds.json').toString());
+const { token, guildid } = JSON.parse(fs.readFileSync('../.creds.json').toString()).archiver;
 const bot = new discord.Client({
     intents: [
         'GuildMessages',
@@ -57,6 +60,7 @@ bot.on('ready', (client) => {
                 if (!fs.existsSync(`../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name}`)) {
                     console.log(`Creating channel folder`)
                     fs.mkdirSync(`../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name}`);
+                    fs.writeFileSync(`../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name}/metadata`, JSON.stringify(channel.toJSON(), null, 4));
                 }
                 try {
                     console.log("Trying 1 fetch operation...");
@@ -71,52 +75,91 @@ bot.on('ready', (client) => {
                     continue;
                 }
                 //otherwise:
-                //make a csv for all messages unless it exists
-                //also check that parent folders exist first
-                if (!fs.existsSync(`../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}`)) {
-                    fs.mkdirSync(`../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}`);
+                //make a db for all messages unless it exists
+                if (!fs.existsSync(`../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name}/messages.sqlite.db`)) {
+                    console.log('Creating message backup database...');
+                    let declaredb = await open({
+                        filename: `../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name}/messages.sqlite.db`,
+                        driver: sqlite3.verbose().Database
+                    });
+                    await declaredb.exec(`CREATE TABLE Messages (
+                        id INTEGER NOT NULL PRIMARY KEY,
+                        content TEXT,
+                        attachments TEXT,
+                        created INTEGER NOT NULL,
+                        edited INTEGER
+                    );`);
+                    await declaredb.close();
                 }
-                if (!fs.existsSync(`../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name}/messages.csv`)) {
-                    console.log('Creating message backup csv...');
-                    fs.writeFileSync(
-                        `../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name}/messages.csv`,
-                        'user,content,attachments,stickers,createdTimestamp,editedTimestamp\n'
-                    );
-                }
-
-                let last: string = 'null';
-                //loop batch message requests (should auto delay itself)
-                console.log('Backing up ' + channel.name);
+                let lastID: null | string = null;
                 let nbatch = 0;
+                const db = await open({
+                    filename: `../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name}/messages.sqlite.db`,
+                    driver: sqlite3.verbose().Database
+                });
+                console.log(`Backing up ${channel.name}`);
                 while (true) {
                     nbatch++;
-                    const fetchopts = (last == 'null') ? { limit: 100 } : { limit: 100, before: last }
-                    let msgs = await channel.messages.fetch(fetchopts);
-                    if (msgs === undefined) {
-                        throw new Error('Message object undefined');
+                    const fetchopts: discord.FetchMessagesOptions = (lastID == null) ? { limit: 100 } : { limit: 100, before: lastID };
+                    const messages = await channel.messages.fetch(fetchopts);
+                    if (channel == null) {
+                        throw new Error('null channel');
                     }
 
-                    console.log(`Backing up ${msgs.size} messages... (batch ${nbatch}: ${msgs.at(-1)?.createdAt.toDateString()} thru ${msgs.at(0)?.createdAt.toDateString()})`);
-                    for (let [_id, msg] of msgs) {
-                        //backup attachment by id
+                    let lastMSG = messages.at(-1);
+                    if (lastMSG == undefined) {
+                        throw new Error('undefined element at -1');
+                    }
+                    lastID = lastMSG.id;
+                    console.log(`Backing up ${messages.size} messages... (batch ${nbatch}, ${messages.at(-1)?.createdAt.toDateString()} thru ${messages.at(0)?.createdAt.toDateString()})`)
+                    for (let [_id, msg] of messages) {
+                        //create array of all attachment and embed ids
+                        let allAttachments = [];
+                        for (let [_id, attachment] of msg.attachments) {
+                            allAttachments.push({ type: 'attachment', id: attachment.id });
+                        }
+                        for (let [_id, sticker] of msg.stickers) {
+                            allAttachments.push({ type: 'sticker', id: sticker.id });
+                        }
+                        for (let embed of msg.embeds) {
+                            allAttachments.push({ type: 'embed', content: embed });
+                        }
+                        //db insert query
+                        await db.exec(`INSERT INTO Messages (
+                                id,
+                                content,
+                                attachments,
+                                created,
+                                edited
+                            ) VALUES (
+                                ${msg.id},
+                                "${Buffer.from(msg.content).toString('base64')}",
+                                "${Buffer.from(JSON.stringify(allAttachments)).toString('base64')}",
+                                ${msg.createdTimestamp},
+                                ${msg.editedTimestamp ? msg.editedTimestamp : 'null'}
+                            )`);
+                        process.stdout.write('🗩');
+
+                        //archive attachments
                         if (msg.attachments.size > 0) {
                             msg.attachments.each((attachment, _key, _) => {
-                                //ignore if it exists
-                                if (!fs.existsSync(`../serverclone/userdata/attachments/${attachment.id}.json`)) {
-                                    // console.log(`Backing up attachment ${attachment.id}`);
+                                if (
+                                    !fs.existsSync(`../serverclone/userdata/attachments/${attachment.id}`) &&
+                                    !fs.existsSync(`../serverclone/userdata/attachments/${attachment.id}.json`)
+                                ) {
+                                    fs.writeFileSync(`../serverclone/userdata/attachments/${attachment.id}.json`, JSON.stringify(attachment, null, 4));
                                     process.stdout.write('🗋');
-                                    fs.writeFileSync(`../serverclone/userdata/attachments/${attachment.id}.json`, JSON.stringify(attachment, null, 4))
                                 }
                             });
                         }
 
-                        //backup sticker by id
-                        if (msg.stickers?.size > 0) {
+                        //archive stickers
+                        if (msg.stickers.size > 0) {
                             msg.stickers.each((sticker, _key, _) => {
-                                //ignore if exists
-                                if (!fs.existsSync(`../serverclone/userdata/stickers/${sticker.id}.json`)) {
-                                    // console.log(`Backing up sticker ${sticker.id}`);
-                                    process.stdout.write('𓉡');
+                                if (
+                                    !fs.existsSync(`../serverclone/userdata/stickers/${sticker.id}`) &&
+                                    !fs.existsSync(`../serverclone/userdata/stickers/${sticker.id}.json`)
+                                ) {
                                     fs.writeFileSync(`../serverclone/userdata/stickers/${sticker.id}.json`, JSON.stringify({
                                         description: sticker.description,
                                         format: sticker.format,
@@ -125,14 +168,16 @@ bot.on('ready', (client) => {
                                         tags: sticker.tags,
                                         url: sticker.url
                                     }, null, 4));
+                                    process.stdout.write('𓉡');
                                 }
-                            });
+                            })
                         }
 
-                        // Write user data if it does not exist
-                        if (!fs.existsSync(`../serverclone/userdata/${msg.author?.id}.json`)) {
-                            // console.log(`\nBacking up user ${msg.author.username}`);
-                            process.stdout.write('𓀀');
+                        //archive users
+                        if (
+                            !fs.existsSync(`../serverclone/userdata/users/${msg.author.id}`) &&
+                            !fs.existsSync(`../serverclone/userdata/users/${msg.author.id}.json`)
+                        ) {
                             fs.writeFileSync(`../serverclone/userdata/${msg.author.id}.json`, JSON.stringify({
                                 name: msg.author.username,
                                 id: msg.author.id,
@@ -140,44 +185,29 @@ bot.on('ready', (client) => {
                                 avatar: msg.author.avatarURL(),
                             }, null, 4));
                         }
-
-                        //create attachment id reference string
-                        let attachids = '';
-                        if (msg.attachments.size > 0) {
-                            msg.attachments.each((attachment, _key, _) => {
-                                attachids += ':' + attachment.id
-                            });
-                        }
-                        //create sticker id reference string
-                        let stickerids = '';
-                        if (msg.stickers.size > 0) {
-                            msg.stickers.each((sticker, _key, _) => {
-                                stickerids += ':' + sticker.id
-                            });
-                        }
-
-                        //write message data to csv
-                        if (channel === null) {
-                            throw new Error('Channel is null');
-                        }
-                        fs.appendFileSync(
-                            `../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name}/messages.csv`,
-                            `${msg.author.id},${encodeURIComponent(msg.content)},${attachids},${stickerids},${msg.createdTimestamp},${msg.editedTimestamp}\n`
-                        );
-                        process.stdout.write('🗩');
                     }
                     process.stdout.write('\n');
-                    if ((msgs.size < 100) || last == msgs.at(-1)?.id) {
-                        console.log(`Under 100 messages (${msgs.size}): ${channel.name} fully backed up`);
+                    if (messages.size < 100) {
+                        console.log(`Under 100 messages (${messages.size}): ${channel.name} fully backed up.`);
                         break;
                     }
-                    let id = msgs.at(-1)?.id
-                    if (id !== undefined) {
-                        last = id;
-                    }
+                }
+                await db.close();
+            }
+            //also clone vc meta
+            else if (channel.type == discord.ChannelType.GuildVoice) {
+                console.log('Saving voice channel metadata...');
+                if (!fs.existsSync(`../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name}`)) {
+                    console.log(`Creating channel folder`)
+                    fs.mkdirSync(`../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name.replace(/\?/gm, '')}`);
+                    fs.writeFileSync(`../serverclone/channels/${(channel.parent) ? channel.parent.name + '/' : ''}${channel.name.replace(/\?/gm, '')}/metadata`, JSON.stringify(channel.toJSON(), null, 4));
                 }
             }
         }
     })
+
+    downloadAllAttachments();
+    downloadAllStickers();
+    downloadAllUsers();
 })
 bot.login(token);
